extracted auth strategies to separate functions
