+++ lib/tasks/import/api.js
@@ -1,9 +1,9 @@
+// @ts-check
 const fs = require('fs-extra');
 const got = require('got');
 const get = require('lodash/get');
 const semver = require('semver');
 const FormData = require('form-data');
-const {Cookie} = require('tough-cookie');
 
 const {promisify} = require('util');
 const stream = require('stream');
@@ -11,6 +11,7 @@ const pipeline = promisify(stream.pipeline);
 const fsp = require('fs').promises;
 
 const {SystemError} = require('../../errors');
+const {v1AuthStrategy, sessionAuthStrategy} = require('./api-auth-strategies.js');
 
 const bases = {
     1: '/ghost/api/v0.1',
@@ -75,59 +76,22 @@ function handleError(error) {
     throw error;
 }
 
+/**
+ * @param {string} version
+ * @param {string} url
+ * @param {object} auth
+ * @param {string} auth.username
+ * @param {string} auth.password
+ */
 async function getAuthOpts(version, url, {username, password}) {
     const baseUrl = getBaseUrl(version, url);
 
-    if (semver.major(version) === 1) {
-        const {body: configBody} = await got('/configuration/', {baseUrl, json: true});
-        const {clientId, clientSecret} = get(configBody, 'configuration[0]', {});
-
-        try {
-            const {body: authBody} = await got.post('/authentication/token/', {
-                baseUrl,
-                json: true,
-                form: true,
-                body: {
-                    grant_type: 'password',
-                    client_id: clientId,
-                    client_secret: clientSecret,
-                    username,
-                    password
-                }
-            });
-
-            return {
-                baseUrl,
-                headers: {
-                    Authorization: `Bearer ${authBody.access_token}`
-                }
-            };
-        } catch (error) {
-            handleError(error);
-        }
-    }
-
     try {
-        const {headers} = await got.post('/session/', {
-            baseUrl,
-            headers: {
-                Origin: url,
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify({username, password})
-        });
+        if (semver.major(version) === 1) {
+            return await v1AuthStrategy(baseUrl, {username, password});
+        }
 
-        /* istanbul ignore next */
-        const cookies = headers['set-cookie'] || [];
-        const filteredCookies = cookies.map(Cookie.parse).filter(Boolean).map(c => c.cookieString());
-
-        return {
-            baseUrl,
-            headers: {
-                Origin: url,
-                Cookie: filteredCookies
-            }
-        };
+        return await sessionAuthStrategy(baseUrl, {origin: url, username, password});
     } catch (error) {
         handleError(error);
     }
